Fix trajopt_ifopt macos build error

+++ tesseract_motion_planners/trajopt_ifopt/src/trajopt_ifopt_motion_planner.cpp
@@ -97,7 +97,7 @@ PlannerResponse TrajOptIfoptMotionPlanner::solve(const PlannerRequest& request) 
   // ----------------
   // Transform plan instructions into trajopt cost and constraints
   std::vector<std::shared_ptr<const trajopt_ifopt::JointPosition>> vars;
-  for (int i = 0; i < move_instructions.size(); ++i)
+  for (int i = 0; i < static_cast<int>(move_instructions.size()); ++i)
   {
     const auto& move_instruction = move_instructions[static_cast<std::size_t>(i)].get().as<MoveInstructionPoly>();
